tests/wpt/web-platform-tests/css/css-toggle/parsing/toggle-group-invalid.tentative.html

<!DOCTYPE HTML>
<meta charset="UTF-8">
<title>CSS Toggles: parsing toggle-group with invalid values</title>
<link rel="author" title="L. David Baron" href="https://dbaron.org/">
<link rel="author" title="Google" href="http://www.google.com/">
<link rel="help" href="https://tabatkins.github.io/css-toggle/#toggle-group-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<body>
<script>

test_invalid_value('toggle-group', 'none, none');
test_invalid_value('toggle-group', 'none self');
test_invalid_value('toggle-group', 'None self');
test_invalid_value('toggle-group', 'NONE self');
test_invalid_value('toggle-group', 'mytoggle self self');
test_invalid_value('toggle-group', 'none, mytoggle');
test_invalid_value('toggle-group', 'mytoggle, none');
test_invalid_value('toggle-group', 'mytoggle 1 self');
test_invalid_value('toggle-group', 'mytoggle self 1');
test_invalid_value('toggle-group', 'mytoggle sticky self');
test_invalid_value('toggle-group', 'mytoggle self sticky');

</script>
